Move university to module window
