+++ src/main/java/com/minecolonies/apiimp/initializer/ModBuildingsInitializer.java
@@ -388,6 +388,7 @@ public final class ModBuildingsInitializer
                                     .setBuildingProducer(BuildingUniversity::new)
                                     .setBuildingViewProducer(() -> BuildingUniversity.View::new)
                                     .setRegistryName(new ResourceLocation(Constants.MOD_ID, ModBuildings.UNIVERSITY_ID))
+                                    .addBuildingModuleViewProducer(() -> UniversityResearchModuleView::new)
                                     .createBuildingEntry();
 
         ModBuildings.hospital = new BuildingEntry.Builder()
