--- src/main/resources/assets/minecolonies/gui/windowhutuniversity.xml
@@ -1,24 +1,0 @@
-<window xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" size="190 244" pause="false"
-        lightbox="false"
-        xsi:noNamespaceSchemaLocation="file:../../../../../api/java/com/minecolonies/blockout/blockOut.xsd">
-    <switch id="pages" size="100% 100%">
-        <view id="pageActions" size="100% 100%">
-            <layout source="minecolonies:gui/layouthuts/layouthutpageactions.xml"/>
-        </view>
-        <view id="pageBranches" size="100% 100%">
-            <image source="minecolonies:textures/gui/builderhut/builder_paper.png" size="190 244"/>
-            <label id="title" size="130 11" pos="30 13" color="black" textalign="MIDDLE" label="$(com.minecolonies.coremod.gui.research.title)"/>
-            <label id="maxresearchwarn" size="180 10" pos="10 28" color="black" textalign="MIDDLE"/>
-            <list id="researches" size="164 175" pos="10 40">
-                <box size="100% 28" linewidth="0">
-                    <buttonimage id="name" size="129 17" pos="2 10" align="MIDDLE" textcolor="black"
-                                 source="minecolonies:textures/gui/builderhut/builder_button_medium_large.png"
-                                 disabled="minecolonies:textures/gui/builderhut/builder_button_medium_large_disabled.png" disabledsize="129 17"/>
-                </box>
-            </list>
-        </view>
-    </switch>
-    <buttonimage id="prevPage" size="18 10" pos="13 13" source="minecolonies:textures/gui/builderhut/turn_page_left.png"/>
-    <buttonimage id="nextPage" size="18 10" pos="159 13" source="minecolonies:textures/gui/builderhut/turn_page_right.png"/>
-    <label id="pageNum" size="20 11" pos="158 222" textalign="MIDDLE" label="" color="black"/>
-</window>
